requirements.txt
coverage==3.6
Django>=1.8,<2.0
django-auth-wall==0.2.0
django-cors-headers==0.12
django-nose==1.4.5
djangorestframework==3.7.7
mock==2.0.0
nose==1.3.7
psycopg2-binary==2.7.4
pycodestyle==2.3.1
Pygments==2.2.0
